Remove unused docker image manager
