+++ pkg/container/images/registry.go
@@ -4,11 +4,17 @@
 package images
 
 import (
+	"archive/tar"
 	"context"
+	"encoding/json"
+	"errors"
 	"fmt"
+	"io"
 	"os"
+	"path/filepath"
 	"runtime"
 
+	"github.com/docker/docker/api/types/build"
 	"github.com/docker/docker/client"
 	"github.com/google/go-containerregistry/pkg/authn"
 	"github.com/google/go-containerregistry/pkg/name"
@@ -129,4 +135,162 @@ func (r *RegistryImageManager) WithKeychain(keychain authn.Keychain) *RegistryIm
 func (r *RegistryImageManager) WithPlatform(platform *v1.Platform) *RegistryImageManager {
 	r.platform = platform
 	return r
+}
+
+// buildDockerImage builds a Docker image using the Docker client API
+func buildDockerImage(ctx context.Context, dockerClient *client.Client, contextDir, imageName string) error {
+	logger.Debugf("Building image %s from context directory %s", imageName, contextDir)
+
+	// Create a tar archive of the context directory
+	tarFile, err := os.CreateTemp("", "docker-build-context-*.tar")
+	if err != nil {
+		return fmt.Errorf("failed to create temporary tar file: %w", err)
+	}
+	defer func() {
+		if err := os.Remove(tarFile.Name()); err != nil {
+			// Non-fatal: temp file cleanup failure
+			logger.Debugf("Failed to remove temporary file %s: %v", tarFile.Name(), err)
+		}
+	}()
+	defer func() {
+		if err := tarFile.Close(); err != nil {
+			// Docker client closes the reader on success, so ignore "already closed" errors
+			if !errors.Is(err, os.ErrClosed) {
+				// Non-fatal: file cleanup failure
+				logger.Debugf("Failed to close tar file: %v", err)
+			}
+		}
+	}()
+
+	// Create a tar archive of the context directory
+	if err := createTarFromDir(contextDir, tarFile); err != nil {
+		return fmt.Errorf("failed to create tar archive: %w", err)
+	}
+
+	// Reset the file pointer to the beginning of the file
+	if _, err := tarFile.Seek(0, 0); err != nil {
+		return fmt.Errorf("failed to reset tar file pointer: %w", err)
+	}
+
+	// Build the image
+	buildOptions := build.ImageBuildOptions{
+		Tags:       []string{imageName},
+		Dockerfile: "Dockerfile",
+		Remove:     true,
+	}
+
+	response, err := dockerClient.ImageBuild(ctx, tarFile, buildOptions)
+	if err != nil {
+		return fmt.Errorf("failed to build image: %w", err)
+	}
+	defer func() {
+		if err := response.Body.Close(); err != nil {
+			// Non-fatal: response body cleanup failure
+			logger.Debugf("Failed to close response body: %v", err)
+		}
+	}()
+
+	// Parse and log the build output
+	if err := parseBuildOutput(response.Body, os.Stdout); err != nil {
+		return fmt.Errorf("failed to process build output: %w", err)
+	}
+
+	return nil
+}
+
+// createTarFromDir creates a tar archive from a directory
+func createTarFromDir(srcDir string, writer io.Writer) error {
+	// Create a new tar writer
+	tw := tar.NewWriter(writer)
+	defer func() {
+		if err := tw.Close(); err != nil {
+			// Non-fatal: tar writer cleanup failure
+			logger.Debugf("Failed to close tar writer: %v", err)
+		}
+	}()
+
+	// Walk through the directory and add files to the tar archive
+	return filepath.Walk(srcDir, func(path string, info os.FileInfo, err error) error {
+		if err != nil {
+			return err
+		}
+
+		// Get the relative path
+		relPath, err := filepath.Rel(srcDir, path)
+		if err != nil {
+			return fmt.Errorf("failed to get relative path: %w", err)
+		}
+
+		// Skip the root directory
+		if relPath == "." {
+			return nil
+		}
+
+		// Create a tar header
+		header, err := tar.FileInfoHeader(info, "")
+		if err != nil {
+			return fmt.Errorf("failed to create tar header: %w", err)
+		}
+
+		// Set the name to the relative path
+		header.Name = relPath
+
+		// Write the header
+		if err := tw.WriteHeader(header); err != nil {
+			return fmt.Errorf("failed to write tar header: %w", err)
+		}
+
+		// If it's a regular file, write the contents
+		if !info.IsDir() {
+			// #nosec G304 - This is safe because we're only opening files within the specified context directory
+			file, err := os.Open(path)
+			if err != nil {
+				return fmt.Errorf("failed to open file: %w", err)
+			}
+			defer func() {
+				if err := file.Close(); err != nil {
+					// Non-fatal: file cleanup failure
+					logger.Debugf("Failed to close file: %v", err)
+				}
+			}()
+
+			if _, err := io.Copy(tw, file); err != nil {
+				return fmt.Errorf("failed to copy file contents: %w", err)
+			}
+		}
+
+		return nil
+	})
+}
+
+// parseBuildOutput parses the Docker image build output and formats it in a more readable way
+func parseBuildOutput(reader io.Reader, writer io.Writer) error {
+	decoder := json.NewDecoder(reader)
+	for {
+		var buildOutput struct {
+			Stream string `json:"stream,omitempty"`
+			Error  string `json:"error,omitempty"`
+		}
+
+		if err := decoder.Decode(&buildOutput); err != nil {
+			if err == io.EOF {
+				break
+			}
+			return fmt.Errorf("failed to decode build output: %w", err)
+		}
+
+		// Check for errors
+		if buildOutput.Error != "" {
+			return fmt.Errorf("build error: %s", buildOutput.Error)
+		}
+
+		// Print the stream output
+		if buildOutput.Stream != "" {
+			if _, err := fmt.Fprint(writer, buildOutput.Stream); err != nil {
+				logger.Debugf("Failed to write build output: %v", err)
+			}
+		}
+	}
+
+	return nil
 }
